Share ImageServerManager
