+++ src/all/batoto/src/eu/kanade/tachiyomi/extension/all/batotov2/BatoToV2.kt
@@ -8,6 +8,7 @@ import android.widget.Toast
 import androidx.preference.CheckBoxPreference
 import androidx.preference.EditTextPreference
 import androidx.preference.PreferenceScreen
+import eu.kanade.tachiyomi.extension.all.batoto.ImageServerManager
 import eu.kanade.tachiyomi.lib.cryptoaes.CryptoAES
 import eu.kanade.tachiyomi.lib.cryptoaes.Deobfuscator
 import eu.kanade.tachiyomi.network.GET
@@ -43,7 +44,6 @@ import java.text.ParseException
 import java.text.SimpleDateFormat
 import java.util.Calendar
 import java.util.Locale
-import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.TimeUnit
 
 open class BatoToV2(
@@ -1138,7 +1138,6 @@ open class BatoToV2(
     }
 
     companion object {
-        private val SERVER_PATTERN = Regex("https://[a-zA-Z]\\d{2}")
         private val seriesUrlRegex = Regex(""".*/series/(\d+)/.*""")
         private val seriesIdRegex = Regex("""series/(\d+)""")
         private val chapterIdRegex = Regex("""/chapter/(\d+)""") // /chapter/4016325
@@ -1157,55 +1156,5 @@ open class BatoToV2(
 
         private val titleRegex: Regex =
             Regex("\\([^()]*\\)|\\{[^{}]*\\}|\\[(?:(?!]).)*]|«[^»]*»|〘[^〙]*〙|「[^」]*」|『[^』]*』|≪[^≫]*≫|﹛[^﹜]*﹜|〖[^〖〗]*〗|\uD81A\uDD0D.+?\uD81A\uDD0D|《[^》]*》|⌜.+?⌝|⟨[^⟩]*⟩|/Official|/ Official", RegexOption.IGNORE_CASE)
-    }
-}
-
-/**
- * Manages image server fallback logic, including blacklisting and backoff tracking.
- */
-private class ImageServerManager() {
-    val serverPattern = Regex("https://([a-zA-Z]\\d{2})")
-
-    val fallbackServers = listOf(
-        "n03", "n00", "n01", "n02", "n04", "n05", "n06", "n07", "n08", "n09", "n10",
-        "k03", "k06", "k00", "k01", "k02", "k04", "k05", "k08", "k09",
-    )
-    val blacklist = listOf("k07")
-
-    // Server status tracking
-    data class ServerStatus(
-        val canBackoff: Boolean,
-        val statusCode: Int = 0,
-        val timestamp: Long = System.currentTimeMillis(),
-    )
-
-    private val serverStatus = ConcurrentHashMap<String, ServerStatus>()
-
-    private val BACKOFF_DURATION_MS = 3_600_000L // 1 hour
-
-    fun shouldSkip(server: String): Boolean {
-        return server in blacklist || isInBackoff(server)
-    }
-
-    fun isInBackoff(server: String): Boolean {
-        val status = serverStatus[server] ?: return false
-        return status.canBackoff && System.currentTimeMillis() - status.timestamp < BACKOFF_DURATION_MS
-    }
-
-    fun recordImageServerStatus(server: String, statusCode: Int) {
-        val now = System.currentTimeMillis()
-        serverStatus[server] = ServerStatus(
-            canBackoff = statusCode in 500..599,
-            statusCode = statusCode,
-            timestamp = now,
-        )
-    }
-
-    fun extractServerFromUrl(url: String): String? {
-        return serverPattern.find(url)?.groups?.get(1)?.value
-    }
-
-    fun replaceServerInUrl(url: String, newServer: String): String {
-        return url.replace(serverPattern, "https://$newServer")
     }
 }
