+++ src/all/batoto/src/eu/kanade/tachiyomi/extension/all/batotov4/BatoToV4.kt
@@ -8,6 +8,7 @@ import android.widget.Toast
 import androidx.preference.CheckBoxPreference
 import androidx.preference.EditTextPreference
 import androidx.preference.PreferenceScreen
+import eu.kanade.tachiyomi.extension.all.batoto.ImageServerManager
 import eu.kanade.tachiyomi.network.GET
 import eu.kanade.tachiyomi.network.POST
 import eu.kanade.tachiyomi.network.asObservableSuccess
@@ -34,7 +35,6 @@ import okhttp3.Response
 import okhttp3.internal.closeQuietly
 import okio.IOException
 import rx.Observable
-import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.TimeUnit
 
 class BatoToV4(
@@ -577,53 +577,3 @@ private const val BROWSE_PAGE_SIZE = 36
 private val seriesIdRegex = Regex("""series/(\d+)""")
 private val titleRegex: Regex =
     Regex("\\([^()]*\\)|\\{[^{}]*\\}|\\[(?:(?!]).)*]|«[^»]*»|〘[^〙]*〙|「[^」]*」|『[^』]*』|≪[^≫]*≫|﹛[^﹜]*﹜|〖[^〖〗]*〗|\uD81A\uDD0D.+?\uD81A\uDD0D|《[^》]*》|⌜.+?⌝|⟨[^⟩]*⟩|/Official|/ Official", RegexOption.IGNORE_CASE)
-
-/**
- * Manages image server fallback logic, including blacklisting and backoff tracking.
- */
-private class ImageServerManager() {
-    val serverPattern = Regex("https://([a-zA-Z]\\d{2})")
-
-    val fallbackServers = listOf(
-        "n03", "n00", "n01", "n02", "n04", "n05", "n06", "n07", "n08", "n09", "n10",
-        "k03", "k06", "k00", "k01", "k02", "k04", "k05", "k08", "k09",
-    )
-    val blacklist = listOf("k07")
-
-    // Server status tracking
-    data class ServerStatus(
-        val canBackoff: Boolean,
-        val statusCode: Int = 0,
-        val timestamp: Long = System.currentTimeMillis(),
-    )
-
-    private val serverStatus = ConcurrentHashMap<String, ServerStatus>()
-
-    private val BACKOFF_DURATION_MS = 3_600_000L // 1 hour
-
-    fun shouldSkip(server: String): Boolean {
-        return server in blacklist || isInBackoff(server)
-    }
-
-    fun isInBackoff(server: String): Boolean {
-        val status = serverStatus[server] ?: return false
-        return status.canBackoff && System.currentTimeMillis() - status.timestamp < BACKOFF_DURATION_MS
-    }
-
-    fun recordImageServerStatus(server: String, statusCode: Int) {
-        val now = System.currentTimeMillis()
-        serverStatus[server] = ServerStatus(
-            canBackoff = statusCode in 500..599,
-            statusCode = statusCode,
-            timestamp = now,
-        )
-    }
-
-    fun extractServerFromUrl(url: String): String? {
-        return serverPattern.find(url)?.groups?.get(1)?.value
-    }
-
-    fun replaceServerInUrl(url: String, newServer: String): String {
-        return url.replace(serverPattern, "https://$newServer")
-    }
-}
